WIP _DataStore - data model refactoring:
At last.
Dropped PickleTankData - under #163
Took the opportunity to bin some more .data (DataDict has a __len__).
It does not blow thanks to short circuiting ShowPanel for classes that
do not have listData. I need a UIListPanel, obviously this must wait for
the Tank panels to go, which in turn depends on #293.