More SelectAndShowItem:
Note that this is a _monkey patch_ - note those calls always follow
RefreshUI. We need a common GetDetailsItem API in panels and then
use that in RefreshUI to avoid repopulating details if it's the
same item (yak) - what the code was trying to do as of old with
`detail='SAME'`. Splitting the Screens/Inis/Installers/People
details classes out stands in the way.

Under #293