Refactoring of file extensions detection API:
File patterns belong to respective classes as much as anything, if not
more. Use class API instead of regexes and bush.game variables directly.

Triggered by @Sharlikran's attempt to extend cosaves files extensions
for fallout. For cosaves:

- Added re.U, does not seem to harm
- Scenarios with renaming/disabling and especially renaming a disabled
save need investigation - we should not have 2 regular expressions but
given that quicksaves and autosaves can not be disabled and renames is
an open issue (#259) this is good enough for now (behavior has not
changed). Note rename for bak files is broken (cosaves are not detected)

Removed outdated comment on LO