Skyrim - handle desync with plugins txt:
Now Bash checks to see if the fetched loadorder.txt is not in sync with
cached_active (must be valid and reflect plugins.txt !). It then
proceeds to reorder the active mods present in lo in the same order as
in plugins.txt and saves the sync'ed loadorder.
For the algorithm see http://stackoverflow.com/q/30504317/281545 - (that
question was motivated by this problem here)

See also a discussion on which file should take priority -
especially the posts by @zilav:
http://forums.bethsoft.com/topic/1599816-wrye-bash-thread-108/page-3#entry25132038
and @Arthmoor:
http://forums.bethsoft.com/topic/1599816-wrye-bash-thread-108/page-4#entry25132514
They both agree that plugins.txt should be given priority.

EDIT: needs testing for get (see '_save_fixed_load_order needs not fetch
the active plugins' commit) and also the algorithm itself needs
rethinking (see next commit).

In getting for skyrim the load order must have already be synced.