Finalizing parent - refactor _toggle_active_state to save only once:
The dialog is rather ugly but it's good for now - it was very important
this was fixed for #210, #256 etc.
As seen in the code _toggle_active_state is only used to either activate
or deactivate - I used this fact to simplify the code but let some of
the flow intact to be able to easily convert this to a real "toggle"
method.