Fixup in duplicate installer (copy_installer):
Hunting down the fomod issue, I run across this one - not sure when this
crept in. The installer.archive attribute was not renamed, so this would
lead to undefined behavior - may be the final root of #93.

Rest is comments and making os_sep unicode just in case (it was str).

Signed-off-by: MrD <the.ubik@gmail.com>