[!]CellPatcher - dropping processing for empty cellData values:
tempCellData will have non empty keys for all cells as at
least one attribute is defined (tags is not empty)
flags may be empty though
The defaultdict really makes a difference in cellData
as this is updated only if the values differ. Before
(on an Import Cells vanilla patch in Skyrim with Update,
Dragonborn and UOP)
cellData would be 5312 entries long most of it empty dicts as
opposed to 253 with the defaultdict (this) patch
Those empty dicts would still result to setCell calls in scanModFile
that seem unneeded (and costly) - unneeded cause in buildPatch
`for attr,value in cellData` will be noop for those empty
dicts so the patch cells were added for no reason.

All started with:

@@ -285,3 +285,4 @@ def checkMasterCellBlockData(cellBlock):
          operator.add,(self.recAttrs[bashKey] for bashKey in tags)))
- except: attrs = set()
+ except TypeError: #reduce() of empty sequence with no initial value
+     attrs = set()
  flags = tuple(self.recFlags[bashKey] for bashKey in tags if

but no processing takes place if attrs is empty so...
There is still something I may not be getting and we need tempCellData
full of empty dicts but I think not

Importers opt:

cachedMasters[srcMod] = srcFile - IIUC comment

Small stuff: Renamed handlePatchCellBlock(patchCellBlock) for clarity