Renamed "Load" menus to "Active":
As there is enough confusion already - "load" should be reserved for
load order related operations - not active mods ones - and selected is
for the selected items in the UI. This will come also
handy when we expose the load orders Bash saves (undo/redo) - we still
need to devise a UI for this.
Added some missing help messages and made sure mods are activated in
load order in "Activate All".
I also edited the readme adding some info on what the code does exactly.

Under #309.