archives refactoring WIP _parse_archive_line:
Archives belong in _bolt_.

Closures and nonlocal:
http://stackoverflow.com/a/19160431/281545
http://stackoverflow.com/q/3190706/281545
https://docs.python.org/3/reference/simple_stmts.html#the-nonlocal-statement

reList did not belong to the installer - it's the standard 7z listing
format, moved to bolt - renamed analogous and unrelated vars
Any cleanup in Installer class variables very welcome - tmp stuff stands
out.

Kept `with tempArch...` in clients as I need the unicode safe path
(whatever that is).

Note that the @balt.conversation does not make a difference - code
returns and by the time I close the log RefreshData is long since
rebound to BashFrame, and triggered. Maybe just leave it that way ?

Under #200 - archives refactoring.