Moving config to gui patchers - WIP needs also next:
Decoupling config from the core patchers.
Note the noop ADoublePatcher is gone - cause it really belongs to the
GUI. Its get/setConfig lacked from _TweakPatcherPanel only:

self.isActive = len(self.enabledTweaks) > 0 ##: NOT HERE !!!!

isActive is set in initPatchFile of the RacePatcher (which is the only
child of DoublePatcher) in the patch phase which runs afterwards, so
no harm

isActive vs isEnabled must be addressed - isActive belongs to the data
model - isEnabled is also used in AMultiTweakItem hierarchy.