Lock load order:
This patch is one of the most important in the 307 dev:

- decouples load_order.py from bosh !
- drops mtimes handling from ModInfos - an Oblivion only concern. In
doing so:
  - simplifies ModInfos refresh from a complex (read ugly) piece of code
  - simplifies RefreshData - always more than welcome
- drops ancient OBMM dialogs and complex dialog logic

Still the error handling in load order needs reworking (under #309)
and we still need to save the load orders on disc (see next commits).
Note that the undo load order feature is essential for user experience.

I initially added a method in BashFrame:

@@ -239,3 +241,3 @@ def get_lo(cached=False, cached_active=True):
             save_lo(saved.loadOrder, saved.activeOrdered)
-            balt.Link.Frame.queue_reset_load_order()
+            warn_locked = True
     return cached_lord

but when modInfos refresh runs for the first time balt.Link.Frame is
None so this blew.

Under #210.