Better undo (skips over invalid till a valid is found) - WIP, readme:
It used to just fail on invalid and user had to rehit Ctrl+Z/Y.
Our redo/undo is a bit peculiar in that when the user makes a change the
new load order/active plugins list is planted in the list of undo/redo
operations instead of binning the redo operations - see
f8634930aff3cecacd3f5d74b03851b9405d5349. I did it this way to persist
as many as possible but still a WIP. Moreover I do not remove invalid
load orders. In 308 we should revisit the whole thing - trying to expose
the persisted orders to the user - we need a lot of thinking on the UI
concerning how to present the active mods vs load order changes.

Under #256, #309.