_wildcard property, support for ghosts and Mod_Prices_Import fix:
Mod_Prices_Import was supposed to be able to import from mods but
the files detection was borged since for ever. Still help items
need revisiting - they report sometimes incorrectly that we can
import from mods while we can and omit to say so when we can

Note I add support for ghosts too.

Signed-off-by: MrD <the.ubik@gmail.com>