Fixup for delete returning in finally - WIP rough edges notes:
In the words of Alex Martelli:
https://www.youtube.com/watch?v=frZrBgWHJdY&feature=youtu.be&t=3m46s

Delete API has still rough edges but this one was probably the next to
last one. One obvious is that because of the non shellUI path we have to
recheck existence of all deleted files in delete_Refresh, although
FileInfos.delete does that. Yak ! Plus if the shell UI path is disabled
(Linux) installers unconditionally call delete_Refresh in DeleteItems.
This last one has to cater for default tweaks, so fix that first (#247).

Under #172.