hspace(pixels):
Idea more or less from:

https://wiki.wxpython.org/BoxSizerFromTheGroundUp/ControlsPadding

It is much more readable and maintainable to add some pixels directly to
the sizer than to OR the RIGHT LEFT TOP BOTTOM flags with the
completely unrelated EXPAND etc, plus eliminates the quasi-magic
numbers. Now all of a sudden all this code makes sense and we can start
factoring out common UI patterns.
Needless to say a super welcome bonus is wx usages outside of balt
dropped by a fair amount.

Initially this included the next commit too, I split it to have the case
where there was a non zero proportion and/or expand flag separately.
Turns out it does not matter:
http://stackoverflow.com/q/38044167/281545 - see especially comment:
http://stackoverflow.com/questions/38044167/wxpython-sizers-does-wx-expand-make-a-difference-in-spacers#comment63692280_38065132
but gives some insight on current sizers patterns.