Dropping the sizer from the main frame:
From: https://wiki.wxpython.org/UsingSizers

"""
If a wxFrame (or derived class) is to have only one child, then you
don't need a sizer at all, as the default behaviour of the frame's
EVT_SIZE handler (assuming there is only one child) is to resize that
child to occupy all of the frame's client area.
"""

No need to return frame for now.