Bash no more creates game.ini without prompt:
saveSettings is called in applyIniTweaks and in two other spots, which
are however not reached if the tweak is invalid (so also if the target
does not exist). Also called in saveSetting which is also guarded.
getBsaRedirection is only called in setBsaRedirection which is also now
guarded if target (game) ini does not exist.
Same goes for applyTweakFile.
This is a WIP: the guards I added belong to the data model, but
editing all the IniFile hierarchy is currently a nono. I need to first
factor out common code (in class methods) and then have apply/save
methods return success/fail info.

Under #290.