Unicode fix in cint Debug_DumpModFiles: EEE pop
The mod u'Blood Raven - ä, ö, ü.esp' displays in the debugger in lo_mods
as:

(9, u'Blood Raven - \xe4, \xf6, \xfc.esp', u'Blood Raven - \xe4, \xf6, \xfc.esp')

and in files as:

u'Load Order (09), Name(Blood Raven - ä, ö, ü.esp)'

This would result in:

Traceback (most recent call last):
  File "bash\balt.py", line 414, in <lambda>
  File "bash\balt.py", line 1558, in _conversation_wrapper
  File "bash\basher\patcher_dialog.py", line 193, in PatchExecute
  File "bash\patcher\patch_files.py", line 631, in buildPatch
  File "bash\game\oblivion\patcher\special.py", line 271, in finishPatch
UnicodeEncodeError: 'ascii' codec can't encode character u'\xe4' in position 449: ordinal not in range(128)

in:

print patchFile.Current.Debug_DumpModFiles() # line 271, in

Monkey patched based on http://stackoverflow.com/a/492711/281545, but as
noted in the comments there this is a monkey patch. Unicode is still an
issue.

Note that it only blows when Bash is run from Wrye Bash Debug.bat - we
may just need to set PYTHONIOENCODING - unfortunately this does not do
it as we redirect to a file and the output encoding prints None. So

@@ -76,2 +76,3 @@ ECHO Launching Wrye Bash 307 in debug mode
 ECHO Launching Wrye Bash 307 in debug mode >>%OUTFILE%
+SET PYTHONIOENCODING=UTF8
 "%PYTHON%" "Wrye Bash Launcher.pyw" -d %1 %2 %3 %4 %5 %6 %7 %8 %9 >>%OUTFILE% 2>&1

Results in:

    print u"input encoding: %s; output encoding: %s; locale: %s" % (
        sys.stdin.encoding,getattr(sys.stdout,'encoding',None),
        locale.getdefaultlocale())

printing:

input encoding: UTF8; output encoding: None; locale: ('en_US', 'cp1253')

due to errLog redirecting the stdout to the bugdumb.

Unfortunately the io module:

    errLog = io.open(os.path.join(os.getcwdu(),u'BashBugDump.log'),'w',encoding='utf-8')

does not work with print - see http://stackoverflow.com/q/14205548/281545
So I fell back to the codecs module which seems to solve the problem.