Unicode fix in cint Debug_DumpModFiles - unicode bugdump:
The mod u'Blood Raven - ä, ö, ü.esp' displays in the debugger in lo_mods
as:

(9, u'Blood Raven - \xe4, \xf6, \xfc.esp', u'Blood Raven - \xe4, \xf6, \xfc.esp')

and in files as:

u'Load Order (09), Name(Blood Raven - ä, ö, ü.esp)'

This would result in:

Traceback (most recent call last):
  File "bash\balt.py", line 414, in <lambda>
  File "bash\balt.py", line 1558, in _conversation_wrapper
  File "bash\basher\patcher_dialog.py", line 193, in PatchExecute
  File "bash\patcher\patch_files.py", line 631, in buildPatch
  File "bash\game\oblivion\patcher\special.py", line 271, in finishPatch
UnicodeEncodeError: 'ascii' codec can't encode character u'\xe4' in position 449: ordinal not in range(128)

in:

print patchFile.Current.Debug_DumpModFiles() # line 271

------------------------------------------------------Mopy/bash/cint.py:
First attempt based on http://stackoverflow.com/a/492711/281545, but as
noted in the comments there this is a monkey patch. Unicode is still an
issue. This worked for this particular case but it's too specific, plus
we'd have to do it all over.

Took the opportunity to wrap a bit in cint - and clarify the code - the
fix is just:

         col.extend(files)
-        return u'\n'.join(col)
+        return u'\n'.join(col).encode('utf-8')

-----------------------------------------------Mopy/Wrye Bash Debug.bat:
I noticed that it only blows when Bash is run from Wrye Bash Debug.bat -
we may just need to set PYTHONIOENCODING ? Nope, this does not do
it as we redirect to a file and the output encoding prints None. So

@@ -76,2 +76,3 @@ ECHO Launching Wrye Bash 307 in debug mode
 ECHO Launching Wrye Bash 307 in debug mode >>%OUTFILE%
+SET PYTHONIOENCODING=UTF8
 "%PYTHON%" "Wrye Bash Launcher.pyw" -d %1 %2 %3 %4 %5 %6 %7 %8 %9 >>%OUTFILE% 2>&1

Results in:

    print u"input encoding: %s; output encoding: %s; locale: %s" % (
        sys.stdin.encoding,getattr(sys.stdout,'encoding',None),
        locale.getdefaultlocale())

printing:

```
input encoding: UTF8; output encoding: None; locale: ('en_US', 'cp1253')
```

due to errLog redirecting the stdout to the bugdumb.
I left it there with a REM that it does not work.

------------------------------------------------------Mopy/bash/bash.py:
Then I thought of solving the problem at the root - set the encoding of
the bugdump.

Unfortunately the io module:

    errLog = io.open(os.path.join(os.getcwdu(),u'BashBugDump.log'),'w',encoding='utf-8')

does not work with print ( http://stackoverflow.com/q/14205548/281545 )
So I fell back to the codecs module which seemed to solve the problem.
Does also solve the Backup settings printing for unicode files:
barb.py  184 Apply: Skyrim\Data\Bash Patches\My_Groups_e???????.csv \
<-- C:\GAMES\Skyrim\Data\Bash Patches\My_Groups_e???????.csv

BUT unfortunately breaks windows newlines - see:
https://bugs.python.org/issue2131
https://github.com/nvaccess/nvda/issues/732
https://github.com/nvaccess/nvda/commit/2d46a35ded41983a9ba1e3e14f2ed252e08ff77a

So bug dump displays in one long line in notepad - bad for the average
user.

--------------------------------------------------------------Final Fix:

codecs.getwriter('utf-8')(open(...))

Yey - see: http://stackoverflow.com/a/40988150/281545

-----------------------------Mopy/bash/game/oblivion/patcher/special.py:
`print _PrintFormID(fid)` would print None to the best of my belief