Pruning getConfigChecked:
Should be moved to the gui side but since it's also used in
initPatchFile we need an abstract override in the data side.
This means that in the MI patchers in gui_patchers.py the gui
patchers must come first - imposing an order in MI is a code
smell however. WIP - allowUnloaded moved to the gui ?
Mopy/bash/patcher/patchers/special.py:
I needed to add self.srcs_ordered cause self.srcs is changed to a set
which results in random order of mods in the log.