Absorb _buildPath in _SimpleImporter:
This allowed me to use it as base in ActorImporter, ImportFactions,
ImportRelations and drop in those (and GraphicsPatcher) the ugly static
decorator from _inner_loop. Threw in a set for faster lookups
(set_id_data) and moved self.classestemp in initPatchFile as it was
there in ActorImporter - not sure about this variable or if should be
indeed a local variable.

id_data = None serves no purpose at all - see next commit.
recAttr dict must be defined in base -> done later