[$]FileInfos.dir -> store_dir:
In one big whoops as otherwise stuff will break.
Actually _DataStore.store_dir - remains to be added.
One Right Way violation (and very annoying at that).
modInfos.dir or self.window.data_store.dir OR dirs['mods']? Still to be
decided. At least now we have a unique searchable name for it.
dir is left as an attribute of FileInfo instances (for now).