[$]Pruning next to last call of getTweakFileSettings from IniInfo:
Had to make _ini_file public - I do not really need to check if the
other file is incompatible, as if incompatible then it is highly
improbable it has the same setting - I do it for the sake of
correctness, it's cheap.
Note fixup in getStatus - pulled `this` (duh) out of the loop and added
test if None - as it was if two inis had no installer (usual case !)
this == other was True resulting in erroneous mismatch = 15. What's
worse is that for every setting, for every ini info it would read all
other ini infos files - meaning m * n^2. Now that's _performance_.
Even now that I use getSettings cache I perform system calls for
checking mtime/size, which would be performed O(n^2) times. I still need
to cache the target ini settings for the getStatus calls.
Mopy/bash/basher/ini_links.py: use cached file instead of rebuilding it

Under #247