Last use of getTweakFileLines - rewrite of listErrors:
Note the bug - it would report 'No valid Batch Script format lines.'
instead of 'Format mismatch:' as of course OBSEIniFile isinstance of
IniFile - the correct way is to use type.
getTweakFileLines is now private (yey !) - its use in getTweakFileLines
should be pruned.
getTweakFileLines is the only method that needs caching now - still used
in few places so we can afford the performance.
Renamed variables in getStatus for clarity.

Under #247