Read bsa header:
Has for the first time in Bash history a test:
https://github.com/wrye-bash/test_bash/commit/aac44daddd0e618389c772881f05663c29b5d91f
The test repo should be a submodule of main repo and commits should
update the submodule but it will be a hassle at this point. So keep
in mind that this branch has tests evolving in parallel.

Was:

def __init__(self):
	self.file_id = 0
	self.version = 0
	self.folder_record_offset = 0
	self.archive_flags = 0
	self.folder_count = 0
	self.file_count = 0
	self.total_folder_name_length = 0
	self.total_file_name_length = 0
	self.file_flags = 0

but evolved ;)