Dropping getAutoItems() from getConfig:
It did not do anything - used to call get_set_choice but due to previous
not needed - if was ever needed. This permitted me to call
_TweakPatcherPanel._log_config in _DoublePatcherPanel - was calling
getConfig which would call getConfig of _ListPanel which would blow
due to executing_patch being None.
getConfig is evil - should be renamed to begin with to "initialize_xxx"
or "load_xxx"