De-wx'ing fonts and display default tweaks in italics:
Incidentally I finally groked why we got the flickering in the mods tab.
If we passed a font in gItem and not _its_ font flickering occurs. So:
So just this:
@@ -1754,5 +1754,5 @@ def __setUI(self, fileName, itemDex):
         if df.back_key: gItem.SetBackgroundColour(colors[df.back_key])
         else: gItem.SetBackgroundColour(self._defaultTextBackground)
-        font = gItem.GetFont()
+        font = df.font
         font.SetUnderlined(df.underline)
         gItem.SetFont(font)
results in flickering! Mystery in #179 solved.

This commit fixes a regression introduced in
da640842f06132e7388953a177107037d23b891c - namely deactivate mods
would not display in italics - not using the Resources font in:
https://github.com/wrye-bash/wrye-bash/commit/da640842f06132e7388953a177107037d23b891c#diff-be6a3eed6c0edcc900fe0291f274970bR866
cured the flickering but the italics were left out.
Why passing the font in as opposed to getting the font from the item
and modify that is worth asking but hey, mystery basically solved after
5 years: https://sourceforge.net/p/oblivionworks/bugs/135/

Under #190.