AMultiTweakItem.tweak_read_classes WIP:
From an engineering point of view, defining a bunch of instance methods
for what is basically class variables just confuses people new to the
code, who keep unfolding methods to discover they basically return a
never changing constant.
This makes apparent that basically the read and write classes for tweaks
are static and makes the cases where they are not stand out.
Not sure why the Cbash patcher needs a list and if basically I can move
the getTypes (rename!) to base and bin getRead/Write classes - which in
the tweaks case just return the same list always.

Note I added a bunch of newlines - so this commit chops off a good 450
lines of unneeded code - and a bunch of "Config Phase" sections, which
should be eliminated from the patchers' data model.

Plus I grow allergic to camelCase APIs.

TODO: investigate rest of the uses, meld getTypes and getRead/Write.

Under #312.