Resizing subwindows on initial run:
To address the points raised by @alt3rn1ty here:
http://forums.bethsoft.com/topic/1612572-wrye-bash-thread-112/#entry25268390
I dropped _sashGravity since it's never used as we always set the sash
position on ShowPanel.
What I really did not fix is why when the panel is not maximized the
bottom/left pane of the details panels minimum size is not respected
Took a note in #197. Tried in mod panel

@@ -1271,3 +1271,5 @@ def __init__(self, parent):
         self._bottom_low_panel.SetSizer(tagsSizer)
-        bottom.SetSizer(vSizer((subSplitter,1,wx.EXPAND)))
+        bSizer = vSizer((subSplitter, 1, wx.EXPAND))
+        bSizer.SetSizeHints(bottom)
+        bottom.SetSizer(bSizer)
         #--Events

but did not work. wx... (plus panels API WIP)