Ini tab refresh fixups and optimizations:
Actually the fixups here is what triggered this branch - cause when the
logic error in refresh() was corrected (the default tweaks were thought
as deleted) the target ini was not read on boot - but that needed
reworking ShowPanel and that in turn needed the ini details panel to be
finally fleshed out.
Dropped `if name in newNames` - we just listed the dir, would not
happen.
Optimization: avoid syscall in getSettings. This also greatly simplifies
needs_update() which before was more like updated(). Further
simplification needed - under #336 essentially.

EDIT: restored the syscall in get_settings:
@@ -1555,3 +1555,3 @@ def getSettings(self, with_deleted=False):
             return ({}, {}) if with_deleted else {}
-        if self._settings_cache_linenum is self.__empty:
+        if self._settings_cache_linenum is self.__empty or self.needs_update():
             self._settings_cache_linenum, self._deleted_cache, \

saveSetting(s) does not have any notion of the cache so I must go
ahead and patch that first - I could tag a cache = __empty in there
but I prefer to leave that for the moment, as saveSettings needs a
rewrite - it should update the cache, not signal it is invalid.
Later - a lot of complicated logic in there.
I did keep `if self._settings_cache_linenum is self.__empty`, on initial
run we get the size/mtime on IniFile.init, no need to restat.