Warn when attempting to drag and drop while list is not sorted by order:
And in the case of mods also when we are trying to move pinned mods.
This leads to confusion which will only get worse when we pin official
DLCs for SSE.
Took the opportunity to rework a bit the ListCtrl API - doDnD was never
set to False and I made fnDndAllow required. There should be one and
only one way...

 # This is the 2nd commit message:

Using different continue keys - we want the user to see both warnings.

 # This is the 3nd commit message:

Fixup for dnd:

Traceback (most recent call last):
  File "bash\basher\__init__.py", line 935, in OnChar
    if not self.dndAllow(event): return
  File "bash\basher\__init__.py", line 817, in dndAllow
    return super(ModList, self).dndAllow(event) # disallow
  File "bash\balt.py", line 1562, in _conversation_wrapper
    return func(*args, **kwargs)
  File "bash\balt.py", line 2116, in dndAllow
    event.Veto()
AttributeError: 'KeyEvent' object has no attribute 'Veto'

@@ -937,3 +937,3 @@ def OnChar(self,event):
         if event.CmdDown() and code in balt.wxArrows:
-            if not self.dndAllow(event): return
+            if not self.dndAllow(event=None): return
             # Calculate continuous chunks of indexes

and:

@@ -2115,3 +2115,3 @@ def PopulateColumns(self):
     def dndAllow(self, event): # Disallow drag an drop by default
-        event.Veto()
+        if event: event.Veto()
         return False