bosh.ini_files:
bosh erosion. Note:

- I kept BestIniFile in bosh as it depends on bosh globals
- I dropped the `if not path:` test - we always pass path in
- reComment = re.compile(u'#.*',re.U) was indeed used in
OBSEIniFile#applyTweakFile. That seems like a bug as the comments in
those files are like in the IniFiles - that is `;.*`. See:
http://forums.bethsoft.com/topic/1612572-wrye-bash-thread-112/page-3#entry25276398
I fixed that:
@@ -580,2 +563,3 @@ def applyTweakFile(self, tweak_lines):
         reDeleted = self.reDeleted
+        reComment = self.reComment
         reSet = self.reSet
Shadowing warnings are there for a reason.
- _ini_types was unused.
- the IDE rearranged the imports - left them that way
- added missing applyTweakFile override in DefaultIniFile.

WIP, but now at least the ini files are in one nice file ready to be
torn to pieces - next goal is to have saveSettings update the cache.

Under #201