Fixup in SelectPage when tabs are disabled:
Took the opportunity to cleanup a bit jump_to_installer API - still we
can't be too smart - the idea would be to check if the installer is
present and then enable the menu but idata is not refreshed till we
ShowPanel for the first time, so it looks like the installer not being
there no matter what. So this:

-        tabInfo[page_title][2].SelectUIListItem(GPath(item),
-                                                deselectOthers=True)
+        target_panel = tabInfo[page_title][2]
+        target_item = GPath(item)
+        if target_item not in target_panel.listData: return
+        target_panel.SelectUIListItem(target_item, deselectOthers=True)

was dropped.

Signed-off-by: MrD <the.ubik@gmail.com>