Fixups for ba2 and minor cleanup - WIP:
The fixups are basically:

- hash are I not Q for BA2s:

@@ -238,3 +238,3 @@ class _B2aFileRecordCommon(_HashedRecord):
     def load_record(self, ins):
-        super(_B2aFileRecordCommon, self).load_record(ins)
+        self.hash, = struct.unpack('I', ins.read(4)) # hash is I not Q !
         for fmt, attr in zip(_B2aFileRecordCommon.formats,

- I never updated "current_folder_name":

@@ -518,2 +519,3 @@ def _load_bsa(self):
                                                              Ba2Folder())
+                current_folder_name = folder_name
             current_folder.assets[file_name] = BSAAsset(file_name,

I also changed the code to store the filename not the full path, to
match behavior on BSAs:

@@ -515,2 +515,3 @@ def _load_bsa(self):
                 folder_name = file_name[:folder_dex]
+                file_name = file_name[folder_dex + 1:]
             if current_folder_name != folder_name:

WIP cause I need to maybe change that to not store the filename twice -
as a dictionary key and an asset attribute. Meanwhile I slotted
BSAAsset.

Some minor cleanup also (using unpacking and common formats handling for
readability, self.filename is unicode - I renamed file_name locals to
easily track setting the filename).