Reworking wx import on boot:
Turns out nothing will function if wx is not imported, due to bosh
importing balt which of course imports wx.
Decoupling bosh from balt is a very laudable goal, but for the moment
we make do: if wx is not present we abort.
We should try in the future restore non gui Bash functionality (or drop
it altogether) - bashmon, print html (are those used ?)
For now, as wx in necessary to run Bash _tinkerSelectGame makes no
sense, hence was dropped.
This commit also drops the balt/basher part of _showErrorInGui -
so we don't display a Bash dialog, just a wx one. KISS, note the
traceback stack lost one level.
I initially had also the bugdump being created for standalone at all
times but I think finally this was a bad idea - it would lead to the
program silently failing from a user's standpoint - I have to revamp
logging anyway but for now I need a bat file.

Under #244