Monkey patch for showing mod details with QUST > GRUP subrecords in fo4:
Report:
http://forums.bethsoft.com/topic/1614408-wrye-bash-thread-113/?p=25292788
Monkey patch in the sense of bush.game.fsName check - we should probably
take care of that in game/ or in the records code. Took the opportunity
to rename a couple local size variables and squash shadowing warnings.
Thanks @zilav for helping me fix this - see discussion starting here:
http://forums.bethsoft.com/topic/1614408-wrye-bash-thread-113/?p=25301448
(lots of info)