[$]Delete API beta (was alpha):
I am finally able to remove delete overrides as well as the
(intentionally) ugly named _delete and itemOrItems etc, etc.
Instead the delete operation is broken down into three overridable
steps: get the paths, delete, refresh. The complexity of delete is
encapsulated in the return values of files_to_delete and input
parameters of delete_refresh (to be renamed to a nice PEP8 name and
be used in refresh soon).
isinstance checks are a big time smell - what if I passed a tuple ?
Or a custom iterable ?
Note the fix for packages that contain the master esm - uninstall would
blow, on trying to delete the master esm - now it will silently skip it,
under #219.

At long last - this finally all but puts to rest #172. Took 3 years