Flip verbose in rescanMergeable:
We want the scan to terminate ASAP in most cases, verbose param was
basically always True (not set ever). Note the test:

@@ -50,5 +51,2 @@ def _is_mergeable_no_load(modInfo, verbose):
     tags = modInfo.getBashTags()
-    if u'NoMerge' in tags:
-        if not verbose: return False
-        reasons.append(u'\n.    '+_(u"Has 'NoMerge' tag."))
     if reasons: return reasons

This was the root of the problem - I could not drop verbose cause the
string (!) was used in rescanMergeable to see if the mod is tagged
NoMerge. This code is unreadable, I have been fighting with it since
forever. At last today I was able to go through all the call graph and
IIUC fix this performance bug - the verbose param being always on,
resulting in 100s Khours of user time lost as the mergeability scan run
before the bashed patch in oblivion. Took the opportunity to factor out
duplicate progress logic - WIP, ofc.