Clear read only flag on plugins txt:
This would lead to crash on start if Bash needed to correct the plugins
txt on start and inability to edit it. See:
http://forums.bethsoft.com/topic/1615914-114/#entry25311868
and ex.
Initial solution I took from:
http://techarttiki.blogspot.gr/2008/08/read-only-windows-files-with-python.html
proved wrong - again same error - it made it into a system file IIUC, so
I used code from bolt - I want this centralized in env however.