BSA conflicts rewrite:
The problem with previous version is it was comparing the order of
a single bsa of the srcInstaller:

    srcBSAOrder = BSAInfos.active_bsa_index(src_bsas[0].name)

(and showInactive was not respected - performance ?)
srcBSAContents became a dict (src_bsa_to_assets).
Now I track every conflict to its src bsa and I respect showLower and
showInactive settings. I should maybe rearrange printing to print more
info. Still a WIP - among others:

- k in _bsas_set and BSAInfos.is_bsa_active(k))
+ k.tail in _bsas_set and BSAInfos.is_bsa_active(k))

is not needed as paths in data_sizeCrc are relative to Data - removed in
next commit.
I now scan the packages in installer order - won't hurt.