Bye _AFileInfo:
Not an accident it had such an ugly name - served its purpose. The rest
is hunting down of FileInfo's attributes that I want to see gone (and
the inevitable renames). This:

-    def __init__(self, dir_, factory=FileInfo):
+    def __init__(self, dir_, factory=AFile):

leads to unresolved attributes warnings in FileInfos that will be
addressed in this merge, when refresh is reworked.

Mopy/bash/basher/misc_links.py: I'd rather avoid using isinstance on
SaveDetails - _append() is always hacky.
Mopy/bash/bosh/_mergeability.py: I don't see why it would not be a mod.

Under #336.