Reworking AFile API - keeping ctime - WIP:
Well - since mtime is reset and size may not change when for instance
flipping the esm bit I hope that the ctime check will catch those.
I still have to go through the uses of setmime and see if it prevents
the crc being recalculated. Note that setting default values in case of
an OSError is now exposed - should be revisited - reset in needs_update
also ?

Note the fix for ghosts - we don't need to update - the statement
`self.ctime = oldCtime` was not doing anything - we would stat and new
ctime would compare different to the old one, so we would update. No
way to set ctime: http://stackoverflow.com/a/5804808/281545. Note also
autoGhost does _not_ run on refresh since a few commits back, I updated
the readme.