_reset_cache rewrite to unify refresh:
And cleanup the API

needs_update and refreshFile reload the cache - that is they readHeader
for Mod/SaveInfo - this is equivalent to what the code was doing but
results in binning the BSAInfos.refresh() override. Getting  there. Note
that Ini/Bsa infos do not reset the cache - the Ini one due to how the
code evolved, while the bsas for performance (memory and speed). The API
is still alpha but getting there - in particular we need a
CorruptedFile(AFile) class to further optimize refresh (now it always
recheck corrupted, plus we do not have an easy way to mark BSAs as
corrupted when their cache is read)

So:

@@ -1423,3 +1423,3 @@ def needs_update(self):
		 if self._file_changed(stat_tuple):
-            self._reset_cache(stat_tuple)
+            self._reset_cache(stat_tuple, load_cache=True)
			 return True
@@ -1430,3 +1430,4 @@ def _file_changed(self, stat_tuple):

and

@@ -2459,5 +2465,5 @@ def refreshFile(self,fileName):
		 try:
-            fileInfo = self.factory(self.store_dir, fileName)
-            fileInfo.readHeader()
+            fileInfo = self.factory(self.store_dir, fileName, load_cache=True)
			 self[fileName] = fileInfo
+            self.corrupted.pop(fileName, None)

I had to move the super constructor call last in FileInfo.init so the
header is not set to None:

@@ -1531,2 +1532,3 @@ def __init__(self, parent_dir, name, load_cache=False):
         self.extras = {}
+        super(FileInfo, self).__init__(self.dir.join(name), load_cache)

But note I short circuit load_cache in BSAInfo#__init__ so I never load
the bsa until its assets are needed.

Also note I now set the mtime/size info to null:

@@ -1421,2 +1421,3 @@ def needs_update(self):
         except OSError:
+            self._reset_cache(self._null_stat, False)
             return False # we should not call needs_update on deleted files

Needs thought.

Under #336.
Under #353.