Slight optimizations in rename - file operations API notes:
Mopy/bash/basher/__init__.py:
No need to self._refresh_detail_info() if we just changed the name - and
no need to backup the bsa info either.
Mopy/bash/bosh/__init__.py: introduced a mark_unchanged() method to
avoid reloading headers for renamed infos (due to ctime change). Turns
out env.shellMove correctly preserves mtime _and_ ctime so this is not
needed - if however we use the shutil methods this is not the case - so
the whole API needs rewrite and needs to be aware of whether we use
SHFileOperation or shutil (an env global). Needless to say this is for
the far future and needs lots of thought. Start by unifying move/copy
info(s) APIs...

Under #336.