Monkey patch for display resolution change crash and getting the icon:
Crash:
http://forums.bethsoft.com/topic/1615914-114/page-3#entry25316850

See #366 for links - when display resolution changes the wxIcon
constructor returns an empty icon. For instance in 125% resolution
I ask for 16x16 I get a 0 size icon - when I specify not the parameters
I get a 40x40 icon - I then proceed to resize (rescale) that.
Note the wx.Icon constructor can take an exe path in and will extract
the icon - path in the form C:\...\path;0 where the part after the
semicolon is the index of the icon in the file.

This is a monkey patch cause I am not familiar with Images API and
cause those image classes are there to "wrap images before wx is
initialized" dunno why.