Centralizing apply ini tweaks logic - WIP:
Needs so much work:

- oblivionIni rename to main_game_ini and encapsulate (move to bass ?)
- OblivionIni may need define its needs_update and base needs_update
raise if file not present
For now I moved the apply tweaks logic in the IniList - note that
if the game ini is deleted the menu item will be disabled so we won't
be offering to create the game ini (in a loop) - except if
allowNewLines is true where the menu item will _not_ be disabled.
I also dropped the self.abs_path.isfile() checks - completely unneeded
and duplicated in saveSettings (I may need to drop them from there too,
let it blow - client should make sure it calls those on existing inis).

Under #326.