Drop Path._csbody:
The only place that csbody was used in a loop was in multitweak_actors,
but that code had performance issues (listing dir and checking existence
twice). So this patch should actually be faster.
Chops off another 16 mbytes in my setup.

 # This is the 2nd commit message:

On a second thought, Path.list returns only filenames in directory so we
don't need the filename body, the filename itself begins with the body,
so I used Path.cs instead.